Temp: Settle helpers, components, globals except timing.h

+++ src/basilisk_teensy/oneshots/_meta.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../servo_units/basilisk.h"
+#include "../basilisk.h"
 
 #define ONESHOT_CRMuxXbee (0)
 #define ONESHOT_SetBaseYaw (1)
@@ -12,13 +12,13 @@ struct BasiliskOneshots {
   static void CRMuxXbee(Basilisk*);
   static void SetBaseYaw(Basilisk*);
 
-  static inline const std::map<uint8_t, void (*)(Basilisk*)> oneshots = {
+  static inline const std::map<int, void (*)(Basilisk*)> oneshots = {
       {ONESHOT_CRMuxXbee, &CRMuxXbee},
       {ONESHOT_SetBaseYaw, &SetBaseYaw},
   };
 
   static void Shoot(Basilisk* b) {
-    for (uint8_t oneshot = 0; oneshot < 8; oneshot++) {
+    for (int oneshot = 0; oneshot < 8; oneshot++) {
       if (b->cmd_.oneshots & (1 << oneshot)) {
         const auto maybe_method = SafeAt(oneshots, oneshot);
         if (maybe_method) {
